requirements.txt
cmake==3.17.2
dlib==19.18.0
cvzone==1.5.6
face_recognition==1.3.0
opencv-python==4.6.0.66
pandas
pyodbc==5.0.1
python-decouple == 3.8
ultralytics==8.0.11
mediapipe





